# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Product API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api-loja-node.onrender.com"
        }
    ],
    "paths": {
        "/products": {
            "get": {
                "summary": "Get all products",
                "responses": {
                    "200": {
                        "description": "A list of products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new product",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "The created product",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/products/update/{id}": {
            "patch": {
                "summary": "Update a product",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "The updated product",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/products/delete/{id}": {
            "delete": {
                "summary": "Delete a product",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The deleted product",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Product": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Camisa"
                    },
                    "description": {
                        "type": "string",
                        "example": "Camisa Boa"
                    },
                    "color": {
                        "type": "string",
                        "example": "Azul"
                    },
                    "weight": {
                        "type": "number",
                        "example": 0.5
                    },
                    "type": {
                        "type": "string",
                        "example": "Vestuario"
                    },
                    "price": {
                        "type": "number",
                        "example": 50.0
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "ProductCreate":
            {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "color": {
                        "type": "string"
                    },
                    "weight": {
                        "type": "number"
                    },
                    "type": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    }
                }
            }
        }
    }
}